+++ src/components/Hero/Hero.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 import { Box, useMultiStyleConfig } from "@chakra-ui/react";
 
 import { HeroTypes, HeroSecondaryTypes } from "./HeroTypes";
+import Image from "../Image/Image";
 
 export interface HeroProps {
   /** Optional hex color value used to override the default background
@@ -20,11 +21,14 @@ export interface HeroProps {
   heading?: JSX.Element;
   /** Used to control how the `Hero` component will be rendered. */
   heroType?: HeroTypes;
-  /** Optional `Image` component used for SECONDARY, FIFTYFIFTY and CAMPAIGN
-   * `Hero` types; Note: `image` can only be used in conjunction with
-   * `backgroundImageSrc` for CAMPAIGN the `Hero` type.
+  /** Text description of the image; to follow best practices for accessibility,
+   * this prop should not be left blank if `imageSrc` is passed. */
+  imageAlt?: string;
+  /** Optional `imageSrc` used for SECONDARY, FIFTYFIFTY and CAMPAIGN
+   * `Hero` types; Note: `imageSrc` can only be used in conjunction with
+   * `backgroundImageSrc` for the CAMPAIGN `Hero` type.
    * Note: not all `Hero` variations utilize this prop. */
-  image?: JSX.Element;
+  imageSrc?: string;
   /** Optional details area that contains location data.
    * Note: not all `Hero` variations utilize this prop. */
   locationDetails?: JSX.Element;
@@ -52,7 +56,8 @@ export default function Hero(props: React.PropsWithChildren<HeroProps>) {
     foregroundColor,
     heading,
     heroType,
-    image,
+    imageAlt,
+    imageSrc,
     locationDetails,
     subHeaderText,
   } = props;
@@ -65,6 +70,12 @@ export default function Hero(props: React.PropsWithChildren<HeroProps>) {
   let backgroundImageStyle = {};
   let contentBoxStyling = {};
 
+  if (imageSrc && !imageAlt) {
+    console.warn(
+      `NYPL Reservoir: The "imageSrc" prop was passed but the "imageAlt" props was not. This will make the rendered image inaccessible.`
+    );
+  }
+
   if (heroType === HeroTypes.Primary) {
     if (!backgroundImageSrc) {
       console.warn(
@@ -72,10 +83,10 @@ export default function Hero(props: React.PropsWithChildren<HeroProps>) {
           "prop for the `HeroTypes.Primary` `heroType` variant."
       );
     }
-    if (image) {
+    if (imageAlt && imageSrc) {
       console.warn(
-        "NYPL Reservoir Hero: The `image` prop has been passed, but the " +
-          "`HeroTypes.Primary` `heroType` variant will not use it."
+        "NYPL Reservoir Hero: The `imageSrc` and `imageAlt` props have been " +
+          "passed, but the `HeroTypes.Primary` `heroType` variant will not use it."
       );
     }
   } else if (locationDetails) {
@@ -90,17 +101,17 @@ export default function Hero(props: React.PropsWithChildren<HeroProps>) {
         "but the `HeroTypes.Secondary` `heroType` variant will not use it."
     );
   }
-  if (heroType === HeroTypes.Tertiary && (backgroundImageSrc || image)) {
+  if (heroType === HeroTypes.Tertiary && (backgroundImageSrc || imageSrc)) {
     console.warn(
       "NYPL Reservoir Hero: The `HeroTypes.Tertiary` `heroType` variant hero " +
         "will not use any of the image props."
     );
   }
-  if (heroType === HeroTypes.Campaign && (!backgroundImageSrc || !image)) {
+  if (heroType === HeroTypes.Campaign && (!backgroundImageSrc || !imageSrc)) {
     console.warn(
       "NYPL Reservoir Hero: It is recommended to use both the " +
-        "`backgroundImageSrc` and `image` props for the `HeroTypes.Campaign` " +
-        "`heroType` variant."
+        "`backgroundImageSrc` and `imageSrc` props for the " +
+        "`HeroTypes.Campaign` `heroType` variant."
     );
   }
   if (heroType === HeroTypes.FiftyFifty && backgroundImageSrc) {
@@ -141,7 +152,7 @@ export default function Hero(props: React.PropsWithChildren<HeroProps>) {
   const childrenToRender =
     heroType === HeroTypes.Campaign ? (
       <>
-        {image}
+        <Image alt={imageAlt} src={imageSrc} />
         <Box __css={styles.interior}>
           {finalHeading}
           {subHeaderText}
@@ -149,9 +160,9 @@ export default function Hero(props: React.PropsWithChildren<HeroProps>) {
       </>
     ) : (
       <>
-        {heroType !== HeroTypes.Primary &&
-          heroType !== HeroTypes.Tertiary &&
-          image}
+        {heroType !== HeroTypes.Primary && heroType !== HeroTypes.Tertiary && (
+          <Image alt={imageAlt} src={imageSrc} />
+        )}
         {finalHeading}
         {heroType === HeroTypes.Tertiary && subHeaderText ? (
           <p>{subHeaderText}</p>