+++ src/index.ts
@@ -17,9 +17,9 @@ export {
   VStack,
 } from "@chakra-ui/react";
 export { default as Accordion } from "./components/Accordion/Accordion";
-export { default as Breadcrumbs } from "./components/Breadcrumbs/Breadcrumbs";
-export { ColorVariants } from "./components/Breadcrumbs/BreadcrumbsTypes";
 export { AccordionTypes } from "./components/Accordion/AccordionTypes";
+export { default as Breadcrumbs } from "./components/Breadcrumbs/Breadcrumbs";
+export { BreadcrumbsTypes } from "./components/Breadcrumbs/BreadcrumbsTypes";
 export { default as Button } from "./components/Button/Button";
 export { ButtonTypes } from "./components/Button/ButtonTypes";
 export {
@@ -28,10 +28,8 @@ export {
   CardContent,
   CardHeading,
 } from "./components/Card/Card";
-export { CardLayouts } from "./components/Card/CardTypes";
 export { default as Checkbox } from "./components/Checkbox/Checkbox";
 export { default as CheckboxGroup } from "./components/CheckboxGroup/CheckboxGroup";
-export { CheckboxGroupLayoutTypes } from "./components/CheckboxGroup/CheckboxGroupLayoutTypes";
 export {
   default as DatePicker,
   FullDateType,
@@ -43,10 +41,7 @@ export { default as Form, FormField, FormRow } from "./components/Form/Form";
 export { FormGaps } from "./components/Form/FormTypes";
 export { GridGaps } from "./components/Grid/GridTypes";
 export { default as Heading } from "./components/Heading/Heading";
-export {
-  HeadingDisplaySizes,
-  HeadingLevels,
-} from "./components/Heading/HeadingTypes";
+export { HeadingSizes, HeadingLevels } from "./components/Heading/HeadingTypes";
 export { default as HelperErrorText } from "./components/HelperErrorText/HelperErrorText";
 export { default as Hero } from "./components/Hero/Hero";
 export { HeroTypes } from "./components/Hero/HeroTypes";
@@ -67,6 +62,7 @@ export {
   ImageTypes,
 } from "./components/Image/ImageTypes";
 export { default as Label } from "./components/Label/Label";
+export { LayoutTypes } from "./helpers/enums";
 export { default as Link } from "./components/Link/Link";
 export { LinkTypes } from "./components/Link/LinkTypes";
 export { default as List } from "./components/List/List";
@@ -84,15 +80,11 @@ export {
 } from "./components/ProgressIndicator/ProgressIndicatorTypes";
 export { default as Radio } from "./components/Radio/Radio";
 export { default as RadioGroup } from "./components/RadioGroup/RadioGroup";
-export { RadioGroupLayoutTypes } from "./components/RadioGroup/RadioGroupLayoutTypes";
 export { default as SearchBar } from "./components/SearchBar/SearchBar";
 export { default as Select } from "./components/Select/Select";
 export { default as SimpleGrid } from "./components/Grid/SimpleGrid";
 export { default as SkeletonLoader } from "./components/SkeletonLoader/SkeletonLoader";
-export {
-  SkeletonLoaderImageRatios,
-  SkeletonLoaderLayouts,
-} from "./components/SkeletonLoader/SkeletonLoaderTypes";
+export { SkeletonLoaderImageRatios } from "./components/SkeletonLoader/SkeletonLoaderTypes";
 export { default as Slider } from "./components/Slider/Slider";
 export { default as StatusBadge } from "./components/StatusBadge/StatusBadge";
 export { StatusBadgeTypes } from "./components/StatusBadge/StatusBadgeTypes";
@@ -118,14 +110,14 @@ export {
   TemplateFooter,
 } from "./components/Template/Template";
 export { default as Text } from "./components/Text/Text";
-export { TextDisplaySizes } from "./components/Text/TextTypes";
+export { TextSizes } from "./components/Text/TextTypes";
 export {
   default as TextInput,
   TextInputRefType,
 } from "./components/TextInput/TextInput";
 export { TextInputTypes } from "./components/TextInput/TextInputTypes";
 export { default as Toggle } from "./components/Toggle/Toggle";
-export { ToggleSizes } from "./components/Toggle/ToggleSizes";
+export { ToggleSizes } from "./components/Toggle/ToggleTypes";
 export { default as useNYPLTheme } from "./hooks/useNYPLTheme";
 export { default as VideoPlayer } from "./components/VideoPlayer/VideoPlayer";
 export {